Create GetAlbum
This script is a web scraper written in Python using the BeautifulSoup library.

The script takes a discogs master release ID as input and retrieves information about that album from the discogs.com website. It extracts the artist name, album title, and release year of the given master release ID.

The script uses the requests library to send HTTP requests to the discogs.com website and BeautifulSoup to parse the HTML content of the response. It then searches for specific elements on the webpage using their class or tag names and extracts the desired information.

The artist name, album title, and release year are printed to the console as output.

Please note that this script requires a valid discogs.com master release ID as input in order to retrieve the desired information.